+++ extensions/preprocess/maintenance-page.php
@@ -1,0 +1,16 @@
+<?php
+
+/**
+ * @file
+ * Preprocess maintenance page function.
+ *
+ * Created by: Topsitemakers
+ * http://www.topsitemakers.com/
+ */
+function ultima_preprocess_maintenance_page(&$vars) {
+  
+  global $user;
+
+  // 
+
+}