+++ CHANGELOG.md
@@ -4,6 +4,16 @@
 
 Initial release.
 
+### October 27, 2012
+
+New features and cleanup:
+
+- Glyphicons styles
+- Maintenance page cleanup and preprocess hook
+- Moved the system templates to the <code>templates/system</code> folder
+- Added body classes based on user roles
+- Added main preprocess hook (<code>template_preprocess()</code>)
+
 <hr>
 
 By: [topsitemakers.com](http://www.topsitemakers.com).